refactor(cluster): extract objects to classes
